requirements.txt
langchain-community>=0.2.0
langchain>=0.2.0
langchain-huggingface>=0.1.0
faiss-cpu>=1.8.0
langchain-google-genai>=1.0.0
sentence-transformers>=2.7.0
python-dotenv>=1.0.0